requirements.txt
requests
Jinja2
beautifulsoup4
openai>=1.84.0
tiktoken
pytest
ghapi
python-dotenv

markdown
streamlit
markdown2
PyYAML>=6.0
